moved codeWord.js and  EventStudent.js Files into user.models.js
